Move actions to their own directory (to make consistent with *coder style)
