+++ easyvvuq/actions/__init__.py
@@ -1,17 +1,5 @@
-# First make sure python version is 3.6+
-import sys
-assert sys.version_info >= (3, 6), (f"Python version must be >= 3.6,"
-                                    f"found {sys.version_info}")
-
-import json
-from .constants import OutputType
-from .campaign import Campaign
-from . import actions
-from . import uqp
-from . import collate
-from . import distributions
-from . import encoders
-from . import decoders
+from .base import BaseAction
+from .execute_local import ExecuteLocal
 
 __copyright__ = """
