--- easyvvuq/default_app_decoders.json
@@ -1,14 +1,0 @@
-{
-  "gauss": {
-    "module_location": "easyvvuq.decoders.gauss",
-    "decoder_name": "GaussDecoder"
-  },
-  "cannon": {
-    "module_location": "easyvvuq.decoders.cannon",
-    "decoder_name": "CannonDecoder"
-  },
-    "csv": {
-    "module_location": "easyvvuq.decoders.simple_csv",
-    "decoder_name": "SimpleCSV"
-  }
-}
