--- easyvvuq/default_app_encoders.json
@@ -1,6 +1,0 @@
-{
-  "generic_template": {
-    "module_location": "easyvvuq.encoders.generic_template",
-    "encoder_name": "GenericEncoder"
-  }
-}
